% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tcpl_based.R
\name{auc_hill_tcpl}
\alias{auc_hill_tcpl}
\title{function to calculate Area Under the Curve (AUC) of the hill model}
\usage{
auc_hill_tcpl(p, lower, upper)
}
\arguments{
\item{p}{a vector containing the Hill model parameters: top, log AC50, hill coefficient}

\item{lower}{lower boundary of x for integration}

\item{upper}{upper boundary of x for integration}
}
\value{
calculated area under the curve (AUC) value
}
\description{
function to calculate Area Under the Curve (AUC) of the hill model
}
