% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tcpl_based.R
\name{hill_model}
\alias{hill_model}
\title{hill mode in ToxCast tcpl package}
\usage{
hill_model(p, x)
}
\arguments{
\item{p}{a vector containing the Hill model parameters: top, log AC50, hill coefficient}

\item{x}{a vector of log concentrations}
}
\value{
calculated y value based on the x and model parameters
}
\description{
hill mode in ToxCast tcpl package
}
