% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tcpl_based.R
\name{log_abs_ec}
\alias{log_abs_ec}
\title{calculate absolute EC_anything based on tcpl hill model}
\usage{
log_abs_ec(p, y)
}
\arguments{
\item{p}{a vector containing the Hill model parameters: top, log AC50, hill coefficient}

\item{y}{the y value}
}
\value{
calculated x value
}
\description{
calculate absolute EC_anything based on tcpl hill model
}
