% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_per_plate}
\alias{normalize_per_plate}
\title{normalize per plate}
\usage{
normalize_per_plate(dt, nctrl = "DMSO")
}
\arguments{
\item{dt}{data.frame contains essential columns with the raw data.}

\item{nctrl}{the name (spid) of the vehicle/solvent control used for calculation}
}
\value{
data.frame with normalized value columns. 'nval_mean' column is the normalized value
calculated using the mean of vehicle control wells, 'nval_median' column is the normalized value
calculated using the median of vehicle control wells.
}
\description{
normalize raw readings as percent of median vehicle control wells
}
\examples{
## normalize demo data
demo_mc_norm <- normalize_per_plate(demo_mc, nctrl = "DMSO")

}
