% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_single_assay}
\alias{normalize_single_assay}
\title{normalize per plate (from single assay data)}
\usage{
normalize_single_assay(dt, nctrl)
}
\arguments{
\item{dt}{data.frame contains essential columns with the raw data.}

\item{nctrl}{the name (spid) of the vehicle/solvent control used for calculation}
}
\value{
data.frame with normalized value columns. 'nval_mean' column is the normalized value
calculated using the mean of vehicle control wells, 'nval_median' column is the normalized value
calculated using the median of vehicle control wells.
}
\description{
normalize raw readings as percent of median/mean vehicle control wells, per assay, per plate.
This function is called by normalize_per_plate, should not be called directly by user.
#'
}
