% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round_df}
\alias{round_df}
\title{round digits of numbers}
\usage{
round_df(df, digits)
}
\arguments{
\item{df}{the dataframe input}

\item{digits}{the specified number of digits}
}
\value{
a dataframe
}
\description{
round numbers in a datafram to specified digits
}
