% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_mc}
\alias{demo_mc}
\title{Example multi-concentration screening data}
\format{See vignette for description of the demo data.}
\usage{
demo_mc
}
\description{
A dataset contains example data for a primary inhibition assay and the parallel cytotoxicity assay
}
\keyword{datasets}
