% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_norm}
\alias{mc_norm}
\title{Normalized example multi-concentration screening data}
\format{See vignette for description of the demo data.}
\usage{
mc_norm
}
\description{
A dataset contains normalized example data for a primary inhibition assay and the parallel cytotoxicity assay
}
\keyword{datasets}
